refacto: publishers (proposal)
